【fix】日志模块完成
修改说明: 日志模块完成
配置更改: init.toml
脚本更改: 无
需求/缺陷来源: 主线版本
平台/产品/模块: 全局
波及分析: 无
自测结果: 自测通过

+++ src/config/init_config.rs
@@ -29,7 +29,7 @@ pub struct DBConfig {
 pub struct LogConfig {
     pub file_level: String,
     pub console_level: String,
-    pub file_max_size: i32,
+    pub file_max_size: u64,
     pub file_split_time: i32
 }
